+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versions since `1.7.0` only track ABI breaks and not API breaks.
   This includes some types from the new API and C++ types such as `std::optional<std::string>` as an alternative to getting a char pointer.
 - `TCODZip` can now load and save paths using `<filesystem>` types.
 - Added `tcod::ImagePtr`.
+- Added functions to manually set context pixel-to-tile conversions in cases where it can't be set automatically.
 
 ## Changes
 - `TCODRandom` is now a movable, non-copyable object.